separate separator to its own class
