+++ src/main/java/Task.java
@@ -1,0 +1,2 @@
+public class Task {
+}
